TravelMate/ml-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0

pydantic==2.5.0
python-dotenv==1.0.0

numpy==1.24.3
pandas==2.1.3
scipy==1.11.4
scikit-learn==1.3.2
joblib==1.3.2

python-multipart==0.0.6
httpx==0.25.1

# Logging / observability (optional but recommended)
loguru==0.7.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

pydantic-settings==2.1.0

google-generativeai>=0.3.0